web: split the OverviewActionBar into its own component
